stl-generator/requirements.txt
numpy>=2.2.5
numpy-stl==2.16.0
